post eval clarity

+++ ex01/src/main.cpp
@@ -6,7 +6,7 @@
 /*   By: rlebigre <rlebigre@student.42angouleme.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/07/14 17:43:44 by rlebigre          #+#    #+#             */
-/*   Updated: 2026/08/18 16:05:39 by rlebigre         ###   ########.fr       */
+/*   Updated: 2026/09/15 18:02:53 by rlebigre         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ void	check_arguments(int argc, char **argv)
 {
 	if (argc != 2)
 		throw "Invalid number of arguments. Put the entire operation in quotes.";
-	if (std::string (argv[0]) != "./RPN")
+	if (std::string(argv[0]) != "./RPN")
 		throw "Invalid program name, should be 'RPN'.";
 }
 
@@ -27,7 +27,7 @@ int	main(int argc, char **argv)
 		check_arguments(argc, argv);
 		if (std::string(argv[1]).empty())
 			return 0;
-		int value = process_line(argv[1]);
+		long value = process_line(argv[1]);
 		std::cout << DBLUE << value << RESET << std::endl;
 	} 
 	catch(const char *e) {
